create custom authentication provider, create jwt filter, update portfolio features with authentication

+++ MyStockTracker/src/main/java/com/wangping/MyStockTracker/service/impl/PortfolioServiceImpl.java
@@ -2,6 +2,7 @@ package com.wangping.MyStockTracker.service.impl;
 
 import com.wangping.MyStockTracker.dto.PortfolioRequestDto;
 import com.wangping.MyStockTracker.dto.PortfolioResponseDto;
+import com.wangping.MyStockTracker.entity.Customer;
 import com.wangping.MyStockTracker.entity.Portfolio;
 import com.wangping.MyStockTracker.entity.Stock;
 import com.wangping.MyStockTracker.repository.PortfolioRepository;
@@ -9,9 +10,14 @@ import com.wangping.MyStockTracker.repository.StockRepository;
 import com.wangping.MyStockTracker.repository.CustomerRepository;
 import com.wangping.MyStockTracker.service.IPortfolioService;
 import lombok.RequiredArgsConstructor;
+import org.apache.coyote.Response;
 import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 
+import javax.sound.sampled.Port;
 import java.time.Instant;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -25,20 +31,23 @@ public class PortfolioServiceImpl implements IPortfolioService {
 
 
     @Override
-    public List<PortfolioResponseDto> getPortfolios() {
-        return portfolioRepository.findAll().stream().map(this::transformToDTO).collect(Collectors.toList());
+    public List<PortfolioResponseDto> getPortfolios(Customer customer) {
+        return portfolioRepository.findByCustomer(customer).stream().map(this::transformToDTO).collect(Collectors.toList());
     }
 
     private PortfolioResponseDto transformToDTO(Portfolio portfolio) {
         PortfolioResponseDto portfolioResponseDto = new PortfolioResponseDto();
-        BeanUtils.copyProperties(portfolio, portfolioResponseDto);
-
+        portfolioResponseDto.setShares(portfolio.getShares());
+        portfolioResponseDto.setPortfolioId(portfolio.getPortfolioId());
         portfolioResponseDto.setSymbol(portfolio.getStock().getSymbol());
+        portfolioResponseDto.setPurchasePrice(portfolio.getPurchasePrice());
+        portfolioResponseDto.setPurchaseDate(portfolio.getPurchaseDate());
+
         return portfolioResponseDto;
     }
 
     @Override
-    public void savePortfolio(PortfolioRequestDto portfolioRequestDto) {
+    public void savePortfolio(PortfolioRequestDto portfolioRequestDto, Customer customer) {
 
         Stock stock = stockRepository.findBySymbol(portfolioRequestDto.getSymbol())
                 .orElseGet(() -> {
@@ -49,17 +58,22 @@ public class PortfolioServiceImpl implements IPortfolioService {
 
         Portfolio portfolio = new Portfolio();
         portfolio.setStock(stock);
+        portfolio.setCustomer(customer);
         portfolio.setShares(portfolioRequestDto.getShares());
         portfolio.setPurchasePrice(portfolioRequestDto.getPurchasePrice());
         portfolio.setPurchaseDate(portfolioRequestDto.getPurchaseDate());
         portfolio.setCreatedAt(Instant.now());
 
         portfolioRepository.save(portfolio);
+
     }
 
     @Override
-    public void deletePortfolio(Long id) {
+    public void deletePortfolio(Long id, Customer customer) {
         Portfolio portfolio = portfolioRepository.findById(id).orElseThrow(()  -> new RuntimeException("Portfolio not found" + id));
+        if (!portfolio.getCustomer().getCustomerId().equals(customer.getCustomerId())) {
+            throw new RuntimeException("Forbidden");
+        }
         portfolioRepository.delete(portfolio);
     }
 }
